Finally flying STABLE again!!!
Added support of acrobatic mode and revived the mode where the stick controls rotation rate (not angle).  Both can be controlled by a receiver channel and are available as compilation flags.

In stable mode, when in rate-control, we got very stable flight with no wobble.  Used the following parameters for pitch/roll:

P: 30
I:  2
D:  0